# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cafe Calories API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/items/{cafe_id}": {
      "post": {
        "tags": [
          "items"
        ],
        "summary": "Add Item",
        "operationId": "add_item_items__cafe_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cafe Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "items"
        ],
        "summary": "List Items",
        "operationId": "list_items_items__cafe_id__get",
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cafe Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemOut"
                  },
                  "title": "Response List Items Items  Cafe Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/items": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "List All Items",
        "description": "List all active menu items across all cafes (for AI recommendations)",
        "operationId": "list_all_items_items_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemOut"
                  },
                  "title": "Response List All Items Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/cart/items": {
      "get": {
        "tags": [
          "cart"
        ],
        "summary": "Get Cart Items",
        "description": "Return detailed cart items for the current user.",
        "operationId": "get_cart_items_cart_items_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {},
                  "type": "array",
                  "title": "Response Get Cart Items Cart Items Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/cart/item/{cart_item_id}": {
      "put": {
        "tags": [
          "cart"
        ],
        "summary": "Update Cart Item",
        "description": "Update quantity (and optionally assignee) for a cart item belonging to the current user.",
        "operationId": "update_cart_item_cart_item__cart_item_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cart_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cart Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cart"
        ],
        "summary": "Delete Cart Item",
        "operationId": "delete_cart_item_cart_item__cart_item_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cart_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cart Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ocr/parse-menu": {
      "post": {
        "tags": [
          "OCR"
        ],
        "summary": "Parse Menu From Pdf",
        "description": "Parse a menu PDF file and extract structured menu items using OCR and Mistral AI.\n\nArgs:\n    file: PDF file containing the menu\n    current_user: Current authenticated user\n    \nReturns:\n    OCRResult containing list of parsed menu items",
        "operationId": "parse_menu_from_pdf_ocr_parse_menu_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_menu_from_pdf_ocr_parse_menu_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/ocr/parse-menu-text": {
      "post": {
        "tags": [
          "OCR"
        ],
        "summary": "Parse Menu From Text",
        "description": "Parse menu text content and extract structured menu items using Mistral AI.\nThis endpoint is useful for testing or when you already have extracted text.\n\nArgs:\n    text_content: Raw text content of the menu\n    current_user: Current authenticated user\n    \nReturns:\n    OCRResult containing list of parsed menu items",
        "operationId": "parse_menu_from_text_ocr_parse_menu_text_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "text_content",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Text Content"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_parse_menu_from_pdf_ocr_parse_menu_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_parse_menu_from_pdf_ocr_parse_menu_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "ingredients": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredients"
          },
          "calories": {
            "type": "integer",
            "title": "Calories"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "servings": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Servings"
          },
          "veg_flag": {
            "type": "boolean",
            "title": "Veg Flag",
            "default": true
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "name",
          "calories",
          "price"
        ],
        "title": "ItemCreate"
      },
      "ItemOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "ingredients": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredients"
          },
          "calories": {
            "type": "integer",
            "title": "Calories"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "servings": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Servings"
          },
          "veg_flag": {
            "type": "boolean",
            "title": "Veg Flag",
            "default": true
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "cafe_id": {
            "type": "integer",
            "title": "Cafe Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "name",
          "calories",
          "price",
          "id",
          "cafe_id",
          "active"
        ],
        "title": "ItemOut"
      },
...
      "OCRMenuItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "calories": {
            "type": "integer",
            "title": "Calories"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "ingredients": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredients"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "servings": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Servings"
          },
          "veg_flag": {
            "type": "boolean",
            "title": "Veg Flag",
            "default": true
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "name",
          "calories",
          "price"
        ],
        "title": "OCRMenuItem"
      },
      "OCRResult": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OCRMenuItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "OCRResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}